aarch64: fix NULL bucket entry crashes and boot to ok> prompt
Five surgical fixes to reach the qualifying pass requirement:

1. dictionary_heat_optimization.c: add `if (!e) continue` guards
   in all three heat-percentile search passes of dict_find_word_heat_aware.
   sf_malloc returns zero-filled arena; new bucket slots are NULL until filled.

2. physics_hotwords_cache.c: same null guard in both the disabled-path
   bucket loop and the fallback bucket search loop.

3. dictionary_management.c (vm_find_word fallback): null guard in the
   manual newest-first scan.

4. dictionary_management.c (sf_fc_reserve): post-realloc memcpy to
   preserve existing entries. Kernel sf_realloc is a bump allocator
   that allocates a fresh zero block without copying — losing all prior
   fc-index pointers when a bucket grows past capacity 8.

5. starforth_words.c (TOP-WORDS): make words_with_heat[] static to avoid
   an 8 KB stack frame that corrupts sp after ExitBootServices; fix %2d
   format specifier (unsupported by kvsnprintf) to plain %d, which was
   misaligning va_args so %.*s read name_len (e.g. 7) as a char* and
   faulted at FAR=0x7.

Vocabulary FC-index also had a bucket null check added in
vocabulary_words.c (vocab_find_word) from an earlier pass.

Result: aarch64 QEMU boots ISO image, runs all POST tests, and reaches
the ok> prompt. VOCABULARY: 7/7 passed. TOP-WORDS: 8/8 passed.
No hardware exceptions.

+++ src/starkernel/capsule/capsule_validate.c
@@ -47,7 +47,7 @@ static const char *validate_result_strings[] = {
 };
 
 const char *capsule_validate_result_str(CapsuleValidateResult result) {
-    if (result < sizeof(validate_result_strings) / sizeof(validate_result_strings[0])) {
+    if ((size_t)result < sizeof(validate_result_strings) / sizeof(validate_result_strings[0])) {
         return validate_result_strings[result];
     }
     return "unknown error";
